db(interests): split in 2 files
