+++ backend/matcha/routes/interests.py
@@ -5,8 +5,6 @@ from flask_jwt_extended import (
     get_jwt_identity,
 )
 
-from matcha.db.user import db_set_interests
-
 from matcha.services.interests import services_interests
 
 bp = Blueprint("interests", __name__)
